Add new APIS for pplnn sdk and misc
