+++ mmdeploy/apis/ncnn/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from mmdeploy.backend.ncnn import is_available, is_plugin_available
+from mmdeploy.backend.ncnn import is_available, is_customops_available
 
-__all__ = ['is_available', 'is_plugin_available']
+__all__ = ['is_available', 'is_customops_available']
 
 if is_available():
     from mmdeploy.backend.ncnn.onnx2ncnn import (get_output_model_file,
