+++ mmdeploy/apis/onnxruntime/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from mmdeploy.backend.onnxruntime import is_available, is_plugin_available
+from mmdeploy.backend.onnxruntime import is_available, is_customops_available
 
-__all__ = ['is_available', 'is_plugin_available']
+__all__ = ['is_available', 'is_customops_available']
