+++ mmdeploy/apis/pplnn/__init__.py
@@ -4,6 +4,10 @@ from mmdeploy.backend.pplnn import is_available
 __all__ = ['is_available']
 
 if is_available():
-    from mmdeploy.backend.pplnn.onnx2pplnn import onnx2pplnn
+    from mmdeploy.backend.pplnn.onnx2pplnn import from_onnx as _from_onnx
+    from ..core import PIPELINE_MANAGER
+    from_onnx = PIPELINE_MANAGER.register_pipeline(
+        func_name='mmdeploy.apis.pplnn.from_onnx')(
+            _from_onnx)
 
-    __all__ += ['onnx2pplnn']
+    __all__ += ['from_onnx']
