+++ mmdeploy/backend/pplnn/__init__.py
@@ -1,6 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import importlib
 
+from .utils import register_engines
+
 
 def is_available():
     """Check whether pplnn is installed.
@@ -11,6 +13,8 @@ def is_available():
     return importlib.util.find_spec('pyppl') is not None
 
 
+__all__ = ['register_engines']
+
 if is_available():
-    from .wrapper import PPLNNWrapper, register_engines
-    __all__ = ['register_engines', 'PPLNNWrapper']
+    from .wrapper import PPLNNWrapper
+    __all__ += ['PPLNNWrapper']
