+++ tests/test_utils/test_util.py
@@ -10,9 +10,9 @@ import pytest
 import torch.multiprocessing as mp
 
 import mmdeploy.utils as util
+from mmdeploy.backend.sdk.export_info import export2SDK
 from mmdeploy.utils import target_wrapper
 from mmdeploy.utils.constants import Backend, Codebase, Task
-from mmdeploy.utils.export_info import dump_info
 from mmdeploy.utils.test import get_random_name
 
 correct_model_path = 'tests/data/srgan.py'
@@ -416,7 +416,7 @@ def test_AdvancedEnum():
     not importlib.util.find_spec('mmedit'), reason='requires mmedit')
 def test_export_info():
     with tempfile.TemporaryDirectory() as dir:
-        dump_info(correct_deploy_cfg, correct_model_cfg, dir, '')
+        export2SDK(correct_deploy_cfg, correct_model_cfg, dir, '')
         deploy_json = os.path.join(dir, 'deploy.json')
         pipeline_json = os.path.join(dir, 'pipeline.json')
         detail_json = os.path.join(dir, 'detail.json')
